requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
langchain
streamlit
transformers
numpy
pandas